+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 * Add plan create/update/find API endpoint
+* Add `TransactionReview` webhook notification
 
 ## 6.4.1
 * Add `exchangeRateQuoteId` to `Transaction.sale`